requirements.txt
# Web Framework
fastapi==0.116.1
uvicorn==0.35.0

# HTTP Client
requests==2.32.3

# OpenAI Integration
openai==0.28.1

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Environment Variables
python-dotenv==1.0.1

# JSON Web Tokens (if needed for authentication)
PyJWT==2.9.0

# Database (if needed)
sqlalchemy==2.0.36

# Testing
pytest==8.3.3
httpx==0.28.1

# Code Quality
black==24.10.0
flake8==7.1.1 